% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lip_simulation.R
\name{simulate_proteins_for_group}
\alias{simulate_proteins_for_group}
\title{Title
#'
Description}
\usage{
simulate_proteins_for_group(
  n_subjects,
  group_id,
  n_reps,
  mask_out,
  pk_missed_prop,
  tr_missed_prop,
  sequences,
  pk_prob = NULL,
  tr_prob = NULL
)
}
\arguments{
\item{$$}{}
}
\description{
Title
#'
Description
}
